devenv: Remove pre-commit.hooks.rustfmt
Somehow it emit this error
```
`cargo metadata` exited with an error: error: rustup could not choose a version of cargo to run, because one wasn't specified explicitly, and no default is configured.
help: run 'rustup default stable' to download the latest stable release of Rust and set it as your default toolchain.
```

https://github.com/cachix/pre-commit-hooks.nix/issues/126